+++ src/main/java/pyre/coloredredstone/items/ItemColoredRedstoneBlock.java
@@ -51,4 +51,9 @@ public class ItemColoredRedstoneBlock extends ItemColored implements IColoredIte
     public Entity createEntity(World world, Entity oldEntityItem, ItemStack itemstack) {
         return createColoredEntityItem(world, oldEntityItem, itemstack);
     }
+
+    @Override
+    public int getItemBurnTime(ItemStack itemStack) {
+        return getBurnTime(itemStack);
+    }
 }