+++ src/main/java/pyre/coloredredstone/items/ItemColoredRedstoneTorch.java
@@ -83,4 +83,9 @@ public class ItemColoredRedstoneTorch extends ItemColored implements IColoredIte
     public Entity createEntity(World world, Entity oldEntityItem, ItemStack itemstack) {
         return createColoredEntityItem(world, oldEntityItem, itemstack);
     }
+
+    @Override
+    public int getItemBurnTime(ItemStack itemStack) {
+        return getBurnTime(itemStack);
+    }
 }