+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.1.0
 ### Added
 - Added Chisel integration for Colored Redstone Blocks (Redstone group).
-- Added option to enable/disable colored names nad tooltips for blocks and items.
+- Added option to enable/disable colored names and tooltips for blocks and items.
+- New colored property - Burnable - brown items and blocks in item form can be used as a fuel. Property can be enabled/disabled in config. In addition, burning time can be set in config (default: 200 ticks - 1 item).
 ## 1.0.1
 ### Added
 - Added translations for color names.